+++ extensions/iceworks-config-helper/src/extension.ts
@@ -1,6 +1,7 @@
 import * as vscode from 'vscode';
 import { connectService, getHtmlForWebview } from '@iceworks/vscode-webview/lib/vscode';
 import { initExtension, registerCommand } from '@iceworks/common-service';
+import { ICEWORKS_ICON_PATH } from '@iceworks/constant';
 import { Recorder } from '@iceworks/recorder';
 import { services, getIsUpdatingJsonFile } from './services';
 import i18n from './i18n';
@@ -63,6 +64,7 @@ export async function activate(context: vscode.ExtensionContext) {
         },
       );
       configWebviewPanel.webview.html = getHtmlForWebview(extensionPath, 'jsonform');
+      configWebviewPanel.iconPath = vscode.Uri.parse(ICEWORKS_ICON_PATH);
       configWebviewPanel.onDidDispose(
         () => {
           configWebviewPanel = undefined;