+++ extensions/iceworks-project-creator/web/src/pages/CustomScaffold/components/ScaffoldForm/index.tsx
@@ -1,5 +1,6 @@
 import React, { useEffect, useRef, useState } from 'react';
-import { Drawer, Icon } from '@alifd/next';
+import classnames from 'classnames';
+import { Icon } from '@alifd/next';
 import LayoutConfig from '../LayoutConfig';
 import { Base64 } from 'js-base64';
 import styles from './index.module.scss';
@@ -7,14 +8,10 @@ import styles from './index.module.scss';
 const Scaffoldform = ({ children, onChange, scaffoldValue }) => {
   const iframeRef = useRef(null);
 
-  const [drawerVisible, setDrawerVisible] = useState<boolean>(false);
+  const [configVisible, setConfigVisible] = useState<boolean>(true);
 
-  const onDrawerOpen = () => {
-    setDrawerVisible(true);
-  };
-
-  const onDrawerClose = () => {
-    setDrawerVisible(false);
+  const onToggle = () => {
+    setConfigVisible(!configVisible);
   };
 
   const sendMessage = (content: string) => {
@@ -35,18 +32,17 @@ const Scaffoldform = ({ children, onChange, scaffoldValue }) => {
     sendMessage(JSON.stringify(scaffoldValue));
   }, [scaffoldValue]);
   return (
-    <div className={styles.scaffoldScaffold}>
+    <div className={styles.scaffoldForm}>
       <div className={styles.content}>
         <iframe ref={iframeRef} className={styles.scaffoldTemplateIframe} frameBorder="0" name="scaffoldTemplate" />
-        <div onClick={() => onDrawerOpen()} className={styles.drawerBtn}><Icon type="set" size="large" /></div>
-        <Drawer
-          visible={drawerVisible}
-          placement="right"
-          width={360}
-          onClose={onDrawerClose}
-        >
-          <LayoutConfig value={scaffoldValue} onChange={onChange} />
-        </Drawer>
+        <div onClick={() => onToggle()} className={classnames(styles.configBtn, { [styles.closeConfigBtn]: configVisible })}><Icon type={configVisible ? 'arrow-right' : 'set'} size="large" /></div>
+        {
+          configVisible && (
+          <div className={styles.config}>
+            <LayoutConfig value={scaffoldValue} onChange={onChange} />
+          </div>
+          )
+        }
       </div>
       <div className={styles.action}>{children}</div>
     </div>