+++ extensions/iceworks-project-creator/webpack.config.js
@@ -1,12 +1,18 @@
 const path = require('path');
+const CopyPlugin = require('copy-webpack-plugin');
 
 const tsConfigPath = path.join(__dirname, 'tsconfig.json');
+const outputPath = path.resolve(__dirname, 'build');
+const scaffoldGeneratorLibPath = path.join(require.resolve('ice-scaffold-generator'), '..');
 
 const config = {
   target: 'node',
   entry: './src/extension.ts',
+  node: {
+    __dirname: false,
+  },
   output: {
-    path: path.resolve(__dirname, 'build'),
+    path: outputPath,
     filename: 'extension.js',
     libraryTarget: 'commonjs2',
     devtoolModuleFilenameTemplate: '../[resource-path]',
@@ -35,6 +41,14 @@ const config = {
       },
     ],
   },
+  plugins: [
+    new CopyPlugin({
+      patterns: [
+        { from: path.join(scaffoldGeneratorLibPath, 'scaffold'), to: path.join(outputPath, 'scaffold') },
+        { from: path.join(scaffoldGeneratorLibPath, 'template'), to: path.join(outputPath, 'template') },
+      ],
+    }),
+  ],
 };
 
 module.exports = (env, argv) => {