+++ extensions/iceworks-app/CHANGELOG.md
@@ -4,6 +4,7 @@
 - refactor: internal optimization
 - fix: preview frame mobile mode switch
 - feat: preview frame add page select
+- feat: auto open preview frame when `npm run start` in O2
 
 ## 0.8.1