+++ extensions/iceworks-app/CHANGELOG.md
@@ -1,8 +1,10 @@
 # Change Log
-
 ## 0.8.2
 
 - refactor: internal optimization
+- fix: preview frame mobile mode switch
+- feat: preview frame add page select
+- feat: auto open preview frame when `npm run start` in O2
 
 ## 0.8.1