+++ extensions/iceworks-app/CHANGELOG.md
@@ -6,6 +6,7 @@
 - fix: preview frame mobile mode switch
 - feat: preview frame add page select
 - feat: auto open preview frame when `npm run start` in O2
+- feat: add custom-scaffold entry
 - chore: add icon for WebviewPanel
 
 ## 0.8.1